backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
sqlalchemy==2.0.23
langchain==0.1.0
openai==1.3.8
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
aiofiles==23.2.1
marshmallow==3.20.1
